BackEnd/api/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
aiosqlite
pydantic
passlib[bcrypt]
python-jose[cryptography]